# On packs.install only register resources from installed pack(s)