--- a/contrib/packs/actions/load.yaml
+++ b/contrib/packs/actions/load.yaml
@@ -9,3 +9,8 @@
       type: "string"
       default: "actions,aliases,sensors"
       description: "Possible options are all, sensors, actions, rules, aliases."
+    packs:
+      type: "array"
+      description: "A list of packs to register / load resources from."
+      items:
+        type: "string"