--- a/contrib/packs/actions/workflows/install.yaml
+++ b/contrib/packs/actions/workflows/install.yaml
@@ -26,11 +26,5 @@
       parameters:
         register: "{{register}}"
         packs: "{{ __results['make a prerun'].result }}"
-      on-success: "restart sensor container"
-    -
-      name: "restart sensor container"
-      ref: "packs.restart_component"
-      parameters:
-        servicename: "st2sensorcontainer"
 
   default: "download pack"