This pull request updates `packs.install` and `packs.register` action to only register resources from packs which are being installed.

Before that, we registered resources from all the packs on install. This could take a very long time in case user had a lot of packs and resources locally. This new change makes the user experience better since the whole process completes faster.

Before:

```bash
(virtualenv)vagrant@vagrant-ubuntu-trusty-64:/data/stanley$ time python ./st2client/st2client/shell.py pack install datadog consul
2016-11-10 10:02:06,123  WARNING - Auth API server is not available, skipping authentication.

	[ succeeded ] download pack
	[ succeeded ] make a prerun
	[ succeeded ] install pack dependencies
	[ succeeded ] register pack
	[  failed   ] restart sensor container

id: 5824459e0640fd53f1cde673
action.ref: packs.install
parameters: 
  packs:
  - datadog
  - consul
status: failed
error: bash: /usr/bin/st2ctl: Permission denied
traceback: None
failed_on: restart sensor container
start_timestamp: 2016-11-10T10:02:06.161971Z
end_timestamp: 2016-11-10T10:03:15.211736Z
+--------------------------+-------------------------+---------------------------+-------------------------+-------------------------------+
| id                       | status                  | task                      | action                  | start_timestamp               |
+--------------------------+-------------------------+---------------------------+-------------------------+-------------------------------+
| 5824459e0640fd53ec23ac12 | succeeded (7s elapsed)  | download pack             | packs.download          | Thu, 10 Nov 2016 10:02:06 UTC |
| 582445a50640fd53ec23ac15 | succeeded (3s elapsed)  | make a prerun             | packs.virtualenv_prerun | Thu, 10 Nov 2016 10:02:13 UTC |
| 582445a90640fd53ec23ac18 | succeeded (14s elapsed) | install pack dependencies | packs.setup_virtualenv  | Thu, 10 Nov 2016 10:02:17 UTC |
| 582445b80640fd53ec23ac1b | succeeded (41s elapsed) | register pack             | packs.load              | Thu, 10 Nov 2016 10:02:32 UTC |
| 582445e10640fd53ec23ac1e | failed (1s elapsed)     | restart sensor container  | packs.restart_component | Thu, 10 Nov 2016 10:03:13 UTC |
+--------------------------+-------------------------+---------------------------+-------------------------+-------------------------------+

real	1m11.403s
user	0m2.357s
```

After:

```bash
(virtualenv)vagrant@vagrant-ubuntu-trusty-64:/data/stanley$ time python ./st2client/st2client/shell.py pack install datadog consul
2016-11-10 10:00:18,205  WARNING - Auth API server is not available, skipping authentication.

	[ succeeded ] download pack
	[ succeeded ] make a prerun
	[ succeeded ] install pack dependencies
	[ succeeded ] register pack
	[  failed   ] restart sensor container

id: 582445320640fd53f1cde670
action.ref: packs.install
parameters: 
  packs:
  - datadog
  - consul
status: failed
error: bash: /usr/bin/st2ctl: Permission denied
traceback: None
failed_on: restart sensor container
start_timestamp: 2016-11-10T10:00:18.227732Z
end_timestamp: 2016-11-10T10:00:46.795525Z
+--------------------------+-------------------------+---------------------------+-------------------------+-------------------------------+
| id                       | status                  | task                      | action                  | start_timestamp               |
+--------------------------+-------------------------+---------------------------+-------------------------+-------------------------------+
| 582445320640fd53ec23ac02 | succeeded (7s elapsed)  | download pack             | packs.download          | Thu, 10 Nov 2016 10:00:18 UTC |
| 582445390640fd53ec23ac05 | succeeded (3s elapsed)  | make a prerun             | packs.virtualenv_prerun | Thu, 10 Nov 2016 10:00:25 UTC |
| 5824453d0640fd53ec23ac08 | succeeded (10s elapsed) | install pack dependencies | packs.setup_virtualenv  | Thu, 10 Nov 2016 10:00:29 UTC |
| 582445480640fd53ec23ac0b | succeeded (4s elapsed)  | register pack             | packs.load              | Thu, 10 Nov 2016 10:00:40 UTC |
| 5824454d0640fd53ec23ac0e | failed (1s elapsed)     | restart sensor container  | packs.restart_component | Thu, 10 Nov 2016 10:00:45 UTC |
+--------------------------+-------------------------+---------------------------+-------------------------+-------------------------------+

real	0m30.134s
user	0m1.921s
sys	0m0.561s
```