@Kami 2016-11-10
Another optimization and improvement - 75597a2dbde887bf9c4ba2b47a3498f8cbaf9b1f

I removed the "restart sensor container" step from the packs.install workflow. For quite a while now, sensor container dynamically polls the database for active sensors and starts them so the sensor container restart is not required anymore after the install.

I did verify that locally, but please correct me if I'm wrong or I am missing something.