@armab 2016-11-10
> This pull request updates packs.install and packs.register action to only register resources from packs which are being installed.

This is great! :+1: x :100: