@enykeev 2016-11-11
This PR should also fix the problem with e2e tests. LGTM, so I gonna go ahead and merge it.